Fixed issue with malformed messages
At the moment if a message is malformed then the mailer crashes. It
looks like this was originally handled in the process_mail method when
it was json (as there's a MultiJson::LoadError rescue in there) but now
the extraction is handled outside that method.

I've moved the catch-all nack into the subscribe block, and am rescuing
an REXML::LoadError in the method that extracts the variables. We
should probably add some tests for malformed messages…

Note: I suspect this code could be made cleaner by re-raising a
specific error, e.g. Mailer::ProcessingFailed, when a custom message is
desired and handling just that and general exceptions in the subscribed
block, rather than having numerous nacks throughout the code.